omited most of open components; added helper script

+++ uninstall_packages_group.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
 
 pkgs_installed=$(pacman -Qg Radeon_Software_for_Linux | cut -f2 -d" ")
-echo Removing: $pkgs_installed
-sudo pacman -Rdd $pkgs_installed
+echo Uninstalling packages of Radeon_Software_for_Linux group...
+if [[ $pkgs_installed != "" ]]; then
+    echo Removing: $pkgs_installed
+    sudo pacman -Rdd $pkgs_installed
+else
+    echo Nothing to uninstall
+fi
 
 # remove cached packages, because they will be considered as currupted after newly created will be added to repo
 echo Removing cached packages
